Create a new JavaScript file for the new functionality
