+++ src/style.css
@@ -25,6 +25,7 @@ h1 {
   padding: 0;
   width: 50%;
   align-items: stretch;
+  max-height: 80%;
   display: flex;
   flex-direction: column;
   box-shadow: 0 0 1rem gray;
@@ -61,6 +62,7 @@ img {
   list-style: none;
   display: flex;
   flex-direction: column;
+  overflow: auto;
 }
 
 li {
@@ -101,6 +103,18 @@ input:checked ~ span {
 
 ::-webkit-input-placeholder {
   font-style: italic;
+}
+
+::-webkit-scrollbar {
+  width: 5px;
+}
+
+::-webkit-scrollbar-track {
+  background-color: #f1f1f1;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: #888;
 }
 
 ::-moz-placeholder {
